pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelliconfig"
version = "0.1.0"
description = "Lightweight LLM serving configuration optimizer for vLLM + llm-d"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "aiconfigurator>=0.8.0",
    "typer>=0.12.0",
    "rich>=13.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.8",
]

[project.scripts]
intelliconfig = "intelliconfig.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"
